+++ src/test/kotlin/com/pestphp/pest/tests/types/fixtures/this/itShortLambdaTest.php
@@ -1,0 +1,3 @@
+<?php
+
+it('has home', fn () => $this-><caret> );