+++ src/test/kotlin/com/pestphp/pest/tests/types/fixtures/thisField/beforeAll.php
@@ -1,0 +1,13 @@
+<?php
+
+class Foo {
+    public $a;
+
+    public function b(){}
+}
+
+\beforeAll(fn() => $this->foo = new Foo());
+
+it('has home', function () {
+    $this->foo-><caret>
+});