+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Basic autocompletion for `$this` for PhpUnit TestCase base class ([#11](https://github.com/pestphp/pest-intellij/pull/11))
 - Line markers now works, but runs the whole file and not the single test. ([#17](https://github.com/pestphp/pest-intellij/pull/17))
 - Add running support with debugger ([#19](https://github.com/pestphp/pest-intellij/pull/19))
+- `$this->field` type support for fields declared in `beforeEach` and `beforeAll` functions ([#22](https://github.com/pestphp/pest-intellij/pull/22))
 
 ## [v0.1.1]
 ### Added